- reorder agent initialization/syncing to fix a race that caused the   testrunner to hang during setup - fix accidental starting of the Akonadi server during shutdown - let the signal handler try harder to shut down the server if the   first attempt was not successful
svn path=/trunk/KDE/kdepimlibs/; revision=923298

+++ tests/testrunner/main.cpp
@@ -35,10 +35,19 @@ static TestRunner *runner = 0;
 void sigHandler( int signal )
 {
   kDebug() << "Received signal" << signal;
-  if ( runner )
-    runner->terminate();
-  if ( setup )
-    setup->shutdown();
+  static int sigCounter = 0;
+  if ( sigCounter == 0 ) { // try clean shutdown
+    if ( runner )
+      runner->terminate();
+    if ( setup )
+      setup->shutdown();
+  } else if ( sigCounter == 1 ) { // force shutdown
+    if ( setup )
+      setup->shutdownHarder();
+  } else { // give up and just exit
+    exit( 255 );
+  }
+  ++sigCounter;
 }
 
 int main( int argc, char **argv )
